plan: align display across plan-family commands; surface CreatedBy
The plan-family commands had drifted from each other in several
small ways since the previous PR. Audit + fixes:

- plan tag get rendered the embedded plan as a thin metadata blurb
  (ID, Steps count, prompt first line, Created) — the shape we
  replaced in plan get during the previous PR. It now embeds the
  full plan body (Inputs / Steps / Outputs sections) using the
  same arrow form plan get uses. Plan-level metadata is omitted:
  ID and SelectionHint already appear in the tag header above; the
  rest is one signadot plan get away. The tag header also gains a
  Selection Hint: line so plan tag list and plan tag get agree on
  what summarises a tag.

  To make this shareable without a circular import (plan imports
  plantag for tag-application after compile/create), the plan-detail
  printer moved into the planshared package as PrintPlanDetails
  alongside a PrintPlanBody variant that skips the top-level header
  for embedded use. plan/printers.go collapses to a thin wrapper.

- Surface PlanStatus.CreatedBy (added in the latest go-sdk drop)
  as a Created By: row in the plan header — used by plan get,
  compile, create, recompile, run, and the embedded view in plan
  tag get (where it renders as Plan Created By: to disambiguate
  against the tag-level Created/Updated fields). Resolution by
  ActorType: user → email, api_key → "api key" + masked id,
  system → "system".

- plan action get's Outputs table no longer carries REQUIRED /
  DEFAULT columns. Those are input concepts that always rendered
  as false / empty for outputs (review feedback on PR #325 that
  hadn't landed). Separate row type with just NAME / SCHEMA.

- planrunnergroup/printers.go: migrate inline time.Parse +
  timeago.NoMax to utils.TimeAgo, matching the cleanup we'd already
  done in plantag and planexec. Drops the time and timeago imports.
